+++ db/ent/product/where.go
@@ -10,47 +10,47 @@ import (
 )
 
 // ID filters vertices based on their ID field.
-func ID(id int) predicate.Product {
+func ID(id int64) predicate.Product {
 	return predicate.Product(sql.FieldEQ(FieldID, id))
 }
 
 // IDEQ applies the EQ predicate on the ID field.
-func IDEQ(id int) predicate.Product {
+func IDEQ(id int64) predicate.Product {
 	return predicate.Product(sql.FieldEQ(FieldID, id))
 }
 
 // IDNEQ applies the NEQ predicate on the ID field.
-func IDNEQ(id int) predicate.Product {
+func IDNEQ(id int64) predicate.Product {
 	return predicate.Product(sql.FieldNEQ(FieldID, id))
 }
 
 // IDIn applies the In predicate on the ID field.
-func IDIn(ids ...int) predicate.Product {
+func IDIn(ids ...int64) predicate.Product {
 	return predicate.Product(sql.FieldIn(FieldID, ids...))
 }
 
 // IDNotIn applies the NotIn predicate on the ID field.
-func IDNotIn(ids ...int) predicate.Product {
+func IDNotIn(ids ...int64) predicate.Product {
 	return predicate.Product(sql.FieldNotIn(FieldID, ids...))
 }
 
 // IDGT applies the GT predicate on the ID field.
-func IDGT(id int) predicate.Product {
+func IDGT(id int64) predicate.Product {
 	return predicate.Product(sql.FieldGT(FieldID, id))
 }
 
 // IDGTE applies the GTE predicate on the ID field.
-func IDGTE(id int) predicate.Product {
+func IDGTE(id int64) predicate.Product {
 	return predicate.Product(sql.FieldGTE(FieldID, id))
 }
 
 // IDLT applies the LT predicate on the ID field.
-func IDLT(id int) predicate.Product {
+func IDLT(id int64) predicate.Product {
 	return predicate.Product(sql.FieldLT(FieldID, id))
 }
 
 // IDLTE applies the LTE predicate on the ID field.
-func IDLTE(id int) predicate.Product {
+func IDLTE(id int64) predicate.Product {
 	return predicate.Product(sql.FieldLTE(FieldID, id))
 }
